requirements.txt
numpy
opencv-python
pandas
Pillow
joblib
